image_classification/requirements.txt
numpy
opencv-python
tqdm
pandas
torch==1.4.0
torchsummary
torchvision==0.5.0
pillow
matplotlib
tensorboard==1.14.0
